Move `zip` util into src
We use it in ToIrListener now, not just in the tests.
