+++ quint/src/util.ts
@@ -13,6 +13,7 @@
  */
 
 import JSONbig from 'json-bigint'
+import lodash from 'lodash'
 
 /** Add this at the end of a switch statement or if/then sequence to enforce exhaustiveness checking
  *
@@ -28,4 +29,18 @@ import JSONbig from 'json-bigint'
  * See https://stackoverflow.com/a/39419171 */
 export function unreachable(object: never): never {
   throw new Error(`impossible: non-exhuastive check should fail during type checking ${JSONbig.stringify(object)}`)
+}
+
+/**  A wrapper around lodash zip that ensures all zipped elements are defined
+ *
+ * Raises `Error` if the arrays are not the same length
+ */
+export function zip<A, B>(a: A[], b: B[]): [A, B][] {
+  return lodash.zip(a, b).map(([x, y]) => {
+    if (x === undefined || y === undefined) {
+      throw new Error('Illegal arguments to zip: array lengths unequal')
+    } else {
+      return [x, y]
+    }
+  })
 }
